\name{pecten.fac}
\alias{pecten.fac}

\title{pecten.fac Dataset}

\description{Dataframe with 1 column describing factors for \code{\link{pecten}} dataset. It gives the names of the gels (as rownames) and the condition for the temperature challenge (15C = control  vs 25C) in the "Condition" column (for details, see Artigaud et al, 2013).}
\references{
Artigaud, S., Gauthier, O. & Pichereau, V. (2013) "Identifying differentially expressed proteins in 2-DE experiments: Inputs from transcriptomics statistical tools." Submitted
}

\author{
Sebastien Artigaud
\email{sebastien.artigaud@gmx.com}
}
\keyword{prot2D}
