\name{pecten}
\alias{pecten}

\title{Pecten Dataset}

\description{Raw Volume data from a 2-DE experiment performed on proteins from \emph{Pecten maximus} gills subjected to a temperature challenge (15C = control  vs 25C). 766 spots were identified with 6 replicates per condition, therefore the dataset is a dataframe of 12 columns and 766 rows (for details, see Artigaud et al, 2013).}
\references{
Artigaud, S., Gauthier, O. & Pichereau, V. (2013) "Identifying differentially expressed proteins in 2-DE experiments: Inputs from transcriptomics statistical tools." Submitted
}

\author{
Sebastien Artigaud
\email{sebastien.artigaud@gmx.com}
}
\keyword{prot2D}